Move chart params
