+++ src/__tests__/assets/chart-params/config.js
