+++ src/__tests__/assets/chart-params/data.js
